Move UserIcon into design components
