+++ src/Components/design/UserIcon/UserIcon.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { PropTypes } from "prop-types";
-import randomElement from "../../Helpers/array/randomElement";
+import randomElement from "../../../Helpers/array/randomElement";
 import "./UserIcon.css";
 
 const BACKGROUND_COLORS = ["#DF0B92", "#097FAA", "#AD6200", "#084391"];
@@ -14,7 +14,7 @@ const randomBackgroundColor = (string) => {
 };
 
 export default function UserIcon(props) {
-  const { firstName, lastName } = props;
+  const { firstName, lastName = "" } = props;
   const backgroundColor = randomBackgroundColor(firstName + lastName);
 
   return (
